dmaengine: stm32-mdma: Add DMA/MDMA chaining support
This patch adds support for M2M transfer triggered by STM32 DMA in order to
transfer data from/to SRAM to/from DDR.

Normally, this mode should not be needed as transferring data from/to DDR
is supported by the STM32 DMA.
However, the STM32 DMA don't have the ability to generate burst transfer
on the DDR as it only embeds only a 4-word FIFO although the minimal burst
length on the DDR is 8 words.
Due to this constraint, the STM32 DMA transfers data from/to DDR in a
single way and could lead to pollute the DDR.
To avoid this, we have to use SRAM for all transfers where STM32 DMA is
involved.

So, we need to add an intermediate M2M transfer handled by the MDMA, which
has the ability to generate burst transfer on the DDR, to copy data
from/to SRAM to/from DDR as described below:
For M2D: DDR --> MDMA --> SRAM --> DMA  --> IP
For D2M: IP  --> DMA  --> SRAM --> MDMA --> DDR

This intermediate transfer is triggered by the STM32 DMA when his transfer
complete flag is set. In that way, we are able to build a DMA/MDMA
chaining transfer completely handled by HW.

This patch clearly adds support for M2M transfer triggered by HW.
This mode is not really available in dmaengine framework as normally M2M
transfers are triggered by SW.

Change-Id: Ib54371803d0d0b295544f0241002b940b0f88555
Change-Id: I1dd63efbd7e5a815e7d2d546a8f863de0ea8d229
Change-Id: I8a0b389a3f26b438f6e1d9009c195bad1bcdea6c
Change-Id: I53da82b437473a1d0f10898aa6ac472fead6b233
Change-Id: Ia0506a0ee4ece3ca4a4e3302343b8d96a364bf33
Change-Id: I71e4e0f131012ee1d0306e2fbbc7be25ba9fae46
Change-Id: Id1e5602ed58af81d541cb413f469198998805575
Change-Id: I11f3658d7399474c395902b5d7c72dfdc05da79a
Change-Id: Ie6b9661cdad56d8efb7a346e459ee59cf08bd456
Upstream-Status: Downstream stm32-gh v5.4-stm32mp [0e1742172d502da59b1bbdd7eca36ba1e69cdecd]
Signed-off-by: M'boumba Cedric Madianga <cedric.madianga@gmail.com>
Signed-off-by: Pierre-Yves MORDRET <pierre-yves.mordret@st.com>
Reviewed-on: https://gerrit.st.com/56040
Reviewed-on: https://gerrit.st.com/87638
Reviewed-on: https://gerrit.st.com/88003
Reviewed-on: https://gerrit.st.com/90374
Reviewed-on: https://gerrit.st.com/89857
Reviewed-on: https://gerrit.st.com/89858
Reviewed-on: https://gerrit.st.com/90312
Reviewed-on: https://gerrit.st.com/102861
Reviewed-by: M'boumba Cedric MADIANGA <cedric.madianga@st.com>
Reviewed-by: Ludovic BARRE <ludovic.barre@st.com>
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@st.com>
Tested-by: M'boumba Cedric MADIANGA <cedric.madianga@st.com>
Tested-by: Amelie DELAUNAY <amelie.delaunay@st.com>