ARM: dts: stm32: Add DRC02 DTO for SoM with populated RSI WiFi
Add DT overlay for the DRC02 board to enable configuration where the
DHSOM inserted into the DRC02 has RSI 9116 WiFi populated on the SoM
and where the microSD slot on the bottom of DRC02 must not be used.

This permits a non-default configuration of the SoM and DRC02 board
used for custom device setup with on-SoM WiFi.

Upstream-Status: Inappropriate
Signed-off-by: Marek Vasut <marex@denx.de>